IO Connections (#4)
* begin adding connection module

* added connection trait and serial port implementation

* updating the communicate trait

* modifying trait and implementation

* applying clippy

* adding back in new method for serial

* adding tcp implementation

* updated module reexports and added unit tests

* cleaned up a couple of things

* making some changes for clippy

* adding dead code trait until it gets used

* more clippy changes

+++ src/config.rs
@@ -8,7 +8,7 @@ use std::path::PathBuf;
 #[serde(tag = "type")]
 pub enum ConnectionType {
     Usb { port: String, baud_rate: u32 },
-    Tcp { address: String, port: u32 },
+    Tcp { address: String, port: u16 },
 }
 
 #[derive(Deserialize)]
@@ -62,9 +62,8 @@ impl Config {
 mod tests {
     use super::*;
     use pretty_assertions::assert_eq;
-    use tempfile::NamedTempFile;
-
     use std::io::Write;
+    use tempfile::NamedTempFile;
 
     #[test]
     fn config_new_fail_empty_file() {
